Refactor and clean up old dispatches
