+++ frontend/src/Components/Filter/CustomFilters/CustomFilter.tsx
@@ -1,5 +1,4 @@
 import React, { useCallback, useEffect, useState } from 'react';
-import { useDispatch } from 'react-redux';
 import IconButton from 'Components/Link/IconButton';
 import SpinnerIconButton from 'Components/Link/SpinnerIconButton';
 import { useDeleteCustomFilter } from 'Filters/useCustomFilters';
@@ -23,7 +22,6 @@ function CustomFilter({
 }: CustomFilterProps) {
   const { deleteCustomFilter, isDeleting, deleteError } =
     useDeleteCustomFilter(id);
-  const dispatch = useDispatch();
   const wasDeleting = usePrevious(isDeleting);
   const [isDeletingInternal, setIsDeletingInternal] = useState(false);
 
@@ -52,7 +50,7 @@ function CustomFilter({
         dispatchSetFilter({ selectedFilterKey: 'all' });
       }
     };
-  }, [isDeletingInternal, dispatchSetFilter, dispatch]);
+  }, [isDeletingInternal, dispatchSetFilter]);
 
   return (
     <div className={styles.customFilter}>
