+++ frontend/src/Components/Form/OAuthInput.tsx
@@ -2,8 +2,8 @@ import React, { useCallback, useEffect } from 'react';
 import SpinnerErrorButton from 'Components/Link/SpinnerErrorButton';
 import { kinds } from 'Helpers/Props';
 import useOAuth from 'OAuth/useOAuth';
-import { getValidationFailures } from 'Store/Selectors/selectSettings';
 import { InputOnChange } from 'typings/inputs';
+import { getValidationFailures } from 'Utilities/selectSettings';
 import { useFormInputGroup } from './FormInputGroupContext';
 
 export interface OAuthInputProps {
