+++ frontend/src/Settings/CustomFormats/CustomFormats/Specifications/EditSpecificationModalContent.tsx
@@ -16,8 +16,8 @@ import { usePendingChangesStore } from 'Helpers/Hooks/usePendingChangesStore';
 import { usePendingFieldsStore } from 'Helpers/Hooks/usePendingFieldsStore';
 import { inputTypes, kinds } from 'Helpers/Props';
 import { useShowAdvancedSettings } from 'Settings/advancedSettingsStore';
-import selectSettings from 'Store/Selectors/selectSettings';
 import { InputChanged } from 'typings/inputs';
+import selectSettings from 'Utilities/selectSettings';
 import translate from 'Utilities/String/translate';
 import { CustomFormatSpecification } from '../useCustomFormats';
 import styles from './EditSpecificationModalContent.css';
