+++ frontend/src/Settings/DownloadClients/RemotePathMappings/EditRemotePathMappingModal.tsx
@@ -1,8 +1,6 @@
-import React, { useCallback } from 'react';
-import { useDispatch } from 'react-redux';
+import React from 'react';
 import Modal from 'Components/Modal/Modal';
 import { sizes } from 'Helpers/Props';
-import { clearPendingChanges } from 'Store/Actions/baseActions';
 import EditRemotePathMappingModalContent, {
   EditRemotePathMappingModalContentProps,
 } from './EditRemotePathMappingModalContent';
@@ -17,18 +15,11 @@ function EditRemotePathMappingModal({
   onModalClose,
   ...otherProps
 }: EditRemotePathMappingModalProps) {
-  const dispatch = useDispatch();
-
-  const handleModalClose = useCallback(() => {
-    dispatch(clearPendingChanges({ section: 'settings.remotePathMappings' }));
-    onModalClose();
-  }, [dispatch, onModalClose]);
-
   return (
-    <Modal size={sizes.MEDIUM} isOpen={isOpen} onModalClose={handleModalClose}>
+    <Modal size={sizes.MEDIUM} isOpen={isOpen} onModalClose={onModalClose}>
       <EditRemotePathMappingModalContent
         {...otherProps}
-        onModalClose={handleModalClose}
+        onModalClose={onModalClose}
       />
     </Modal>
   );
