+++ frontend/src/Settings/Metadata/Metadata/EditMetadataModal.tsx
@@ -1,9 +1,7 @@
-import React, { useCallback } from 'react';
-import { useDispatch } from 'react-redux';
+import React from 'react';
 import Modal from 'Components/Modal/Modal';
 import { sizes } from 'Helpers/Props';
 import { useShowAdvancedSettings } from 'Settings/advancedSettingsStore';
-import { clearPendingChanges } from 'Store/Actions/baseActions';
 import EditMetadataModalContent, {
   EditMetadataModalContentProps,
 } from './EditMetadataModalContent';
@@ -18,21 +16,14 @@ function EditMetadataModal({
   onModalClose,
   ...otherProps
 }: EditMetadataModalProps) {
-  const dispatch = useDispatch();
-
   const advancedSettings = useShowAdvancedSettings();
 
-  const handleModalClose = useCallback(() => {
-    dispatch(clearPendingChanges({ section: 'metadata' }));
-    onModalClose();
-  }, [dispatch, onModalClose]);
-
   return (
-    <Modal size={sizes.MEDIUM} isOpen={isOpen} onModalClose={handleModalClose}>
+    <Modal size={sizes.MEDIUM} isOpen={isOpen} onModalClose={onModalClose}>
       <EditMetadataModalContent
         {...otherProps}
         advancedSettings={advancedSettings}
-        onModalClose={handleModalClose}
+        onModalClose={onModalClose}
       />
     </Modal>
   );
