+++ frontend/src/Settings/Tags/AutoTagging/Specifications/EditSpecificationModalContent.tsx
@@ -16,8 +16,8 @@ import { usePendingChangesStore } from 'Helpers/Hooks/usePendingChangesStore';
 import { usePendingFieldsStore } from 'Helpers/Hooks/usePendingFieldsStore';
 import { inputTypes, kinds } from 'Helpers/Props';
 import { useShowAdvancedSettings } from 'Settings/advancedSettingsStore';
-import selectSettings from 'Store/Selectors/selectSettings';
 import { InputChanged } from 'typings/inputs';
+import selectSettings from 'Utilities/selectSettings';
 import translate from 'Utilities/String/translate';
 import { AutoTaggingSpecification } from '../useAutoTaggings';
 import styles from './EditSpecificationModalContent.css';
