+++ frontend/src/Utilities/selectSettings.ts
